Add test for missing objects
This patch adds tests for using a datasource to read an object that does
not exist. When this happen a nice error message should be returned to
the user.

Some endpoints return a 403 or 500 error when this happens, it will need
to be fixed later but for now I'm working around it.

Also adds a user agent string to the client.

Also automatically saves the result to the state by using introspection
and struct tags.