Implement activity support
This patch adds support for the Activity module in the client.
Methods that receive an activity ID from the API now return it to the
user and a WaitForCompletion() can be used to make sure that the
activity finished properly.

Two new `cloudtemple_activity` and `cloudtemple_activities` datasources
have been added to the provider.

The `cloudtemple_virtual_machine` resource now properly check for the
completion of the activities rather than just sleeping a fixed amount of
time.